### 2024-10-10 #7345 @warrickball
Hi @matteobachetti, and thanks for your submission. I noticed that this submission is for Stingray v2 and Stingray v1 was [previously published](https://joss.theoj.org/papers/10.21105/joss.01393) as a joint publication with AAS Journals. Is this submission also a joint submission with AAS Journals?

Also, for the sake of whichever editor and reviewers handle this, are there specific parts of the package that are new and warrant review, or has enough changed that it's worth treating this as a fresh review anyway?

I think the error in the paper is that John Swinbank has affiliation number 4 but no such affiliation is given. ([The list skips from 3 to 5](https://github.com/StingraySoftware/stingray/blob/f3e7d457b037c9688eae7580c0ded0592fc540ba/joss/paper.md?plain=1#L54-L80).)

Note also that we require a statement of need, and that the paper is currently a bit longer than our [stated maximum](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) of 1000 words (though not by much). If you could squeeze it down, that'll be great.

Our editors are all operating at capacity so I'm going to waitlist this submission but hopefully one will be available by the time you've addressed these initial points.
