### 2024-10-10 #7345 @matteobachetti
@warrickball thanks! First of all, let me apologize for not specifying that the paper already went through pyOpenSci review here: https://github.com/pyOpenSci/software-submission/issues/201 (at pyOpenSci the editors told us not to forget this, but of course I did).

There is no publication with AAS planned for now, this paper was really meant to give an update on the work we've been doing on the code since 2019, in connection with the pyOpenSci review.
