### 2024-10-11 #7345 @arfon
@matteobachetti – it would be useful to have an explicit diff here that clearly demonstrates the code changes between v1 and v2 of this library.

@warrickball – assuming the scale/scope of the changes between v1 and v2 meet out 'substantial scholarly effort' criterion, we can proceed here without additional review (see https://joss.readthedocs.io/en/latest/editing.html#ropensci-reviewed-or-pyopensci-reviewed-and-accepted-submissions)
