### 2024-10-11 #7345 @matteobachetti
@arfon, the first paper was in 2019, so basically everything you see in this changelog from 2020 on (from v.0.2 on!) was not in the first paper:
https://docs.stingray.science/en/stable/history.html
