### 2024-10-15 #7345 @matteobachetti
@warrickball I still have a few words to cut (and if you ask how many, it's 42, Douglas Adams would be proud).
However, I get an error from the pdf generation which I don't understand:
```Run bundle exec ruby /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb
/opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych/parser.rb:62:in `_native_parse': (tmp/89fbb0b4faffadd5c23d73ca35df12df/joss/paper.md): did not find expected key while parsing a block mapping at line 2 column 1 (Psych::SyntaxError)
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych/parser.rb:62:in `parse'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:455:in `parse_stream'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:399:in `parse'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:323:in `safe_load'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:369:in `load'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:671:in `block in load_file'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:670:in `open'
	from /opt/hostedtoolcache/Ruby/3.3.4/x64/lib/ruby/3.3.0/psych.rb:670:in `load_file'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:113:in `load_metadata'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:20:in `initialize'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:82:in `new'
	from /home/runner/work/_actions/xuanxu/paper-action/main/vendor/bundle/ruby/3.3.0/gems/theoj-1.10.0/lib/theoj/paper.rb:82:in `from_repo'
	from /home/runner/work/_actions/xuanxu/paper-action/main/get_paper.rb:12:in `<main>'
Error: Process completed with exit code 1.
```

I don't understand what key it is missing in line 2 column 1 🤔
