### 2024-10-15 #7345 @warrickball
It looks to me like [this commit](https://github.com/StingraySoftware/stingray/commit/096eb7dc84b1d52a1e6b24c09a17847e40e8c5ed) nudged the indentation of the `index` field for each affiliation by removing one leading space. I couldn't build the paper as it is but aligning them with the `name` line above again got the paper to build for me.
