### 2024-10-16 #7345 @warrickball
> @matteobachetti – it would be useful to have an explicit diff here that clearly demonstrates the code changes between v1 and v2 of this library.
>
> @warrickball – assuming the scale/scope of the changes between v1 and v2 meet out 'substantial scholarly effort' criterion, we can proceed here without additional review (see https://joss.readthedocs.io/en/latest/editing.html#ropensci-reviewed-or-pyopensci-reviewed-and-accepted-submissions)

```
$ git diff --stat v0.2 v2.0.0 stingray/*.py
 stingray/__init__.py           |    7 +
 stingray/_astropy_init.py      |   35 +--
 stingray/base.py               | 2828 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 stingray/bexvar.py             |  275 ++++++++++++++++++++
 stingray/bispectrum.py         |  105 ++++----
 stingray/conftest.py           |   34 +--
 stingray/covariancespectrum.py |  170 ++++++------
 stingray/crosscorrelation.py   |  175 +++++++++----
 stingray/crossspectrum.py      | 3031 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++----------------------------------------------------
 stingray/events.py             |  816 +++++++++++++++++++++++++++++++++++++---------------------
 stingray/filters.py            |   54 ++--
 stingray/fourier.py            | 2474 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 stingray/gti.py                | 1186 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++---------------------
 stingray/io.py                 | 1393 +++++++++++++++++++++++++++++++++++++++++++++++++++-----------------------------------------------
 stingray/lightcurve.py         | 1446 +++++++++++++++++++++++++++++++++++++++++++++++++++++++-----------------------------------------------
 stingray/loggingconfig.py      |   39 +++
 stingray/lombscargle.py        |  686 +++++++++++++++++++++++++++++++++++++++++++++++++
 stingray/multitaper.py         |  791 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 stingray/power_colors.py       |  495 +++++++++++++++++++++++++++++++++++
 stingray/powerspectrum.py      | 1570 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++------------------------
 stingray/sampledata.py         |   13 +-
 stingray/spectroscopy.py       |   57 ++---
 stingray/stats.py              |  833 +++++++++++++++++++++++++++++++++++++++++++++++++++++++----
 stingray/utils.py              | 1681 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++------------
 stingray/varenergyspectrum.py  |  964 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++------------
 25 files changed, 17514 insertions(+), 3644 deletions(-)
```

@arfon — This isn't necessarily super accurate but I couldn't get `cloc` to parse a `diff`. From the code alone it looks like plenty of functionality was implemented since the first Stingray JOSS paper.
