### 2024-10-16 #7345 @warrickball
@matteobachetti — I intend to take a closer look at this soon but at a quick glance I noticed a few references aren't rendering. Specifically, note all the lingering square brackets in this sentence:

> Methods include Gaussian Process modeling of quasi-periodic oscillations [hubner] and Lomb-Scargle cross spectra [scargle]. We have introduced the Fourier-Domain Acceleration Search [ransom], the H-test [dejager] and Phase Dispersion Minimization [stellingwerf] statistics into the pulsar sub package.

I also don't see the JOSS reference to the original Stingray publication (Huppenkothen et al. 2019b) actually turn up in the text. The associated ApJ reference (Huppenkothen et al. 2019a) is in the summary.

Another few quick fixes while we're at it:
* affiliations 8 & 11 need a country,
* affiliation 13 should be "USA" rather than "US", and
* the first reference (to Astropy Collaboration et al.) should be in brackets.

Don't worry too much about the typographic stuff. I can open a pull request to fix any minor things (like the last point) but it'll be more difficult for me to guess missing data, like exactly what you meant to cite in that sentence about methods that I quoted above.
