### 2024-10-17 #7345 @matteobachetti
@warrickball sorry, I had got wrong in various creative ways the syntax for citations 😅 . Now all should appear correctly
