### 2024-10-21 #7389 @editorialbot
[REVIEW]: Stingray 2: A fast and modern Python library for spectral timing
**Submitting author:** @matteobachetti (<a href="http://orcid.org/0000-0002-4576-9337">Matteo Bachetti</a>)
**Repository:** https://github.com/stingraysoftware/stingray
**Branch with paper.md** (empty if default branch): update_joss_paper
**Version:** v2.2
