### 2024-10-21 #7389 @warrickball
This submission has already been reviewed by pyOpenSci: https://github.com/pyOpenSci/software-submission/issues/201
