### 2024-10-21 #7389 @matteobachetti
@warrickball please note that the version that received the review is 2.1, but the version including the changes made as part of the pyOpenSci review is 2.2 (which will be released tomorrow).
