### 2024-10-22 #7389 @matteobachetti
Here is the new release: https://github.com/StingraySoftware/stingray/releases/tag/v2.2
