### 2024-10-23 #7389 @warrickball
I've opened a [pull request](https://github.com/StingraySoftware/stingray/pull/854) with some editorial tweaks. Since I'm already thinking in AEiC mode, this should be the only pass, unless I need to fix any metadata that's generated when we do a dry run of the publication process.

I did also notice that the Zenodo archive doesn't include ORCIDs for any of the authors. That isn't a requirement for JOSS publication but you might want to add it. If you publish to Zenodo via GitHub integration, it might automatically pick up ORCIDs for those contributors who've added ORCID information to their GitHub profile.
