### 2024-10-24 #7389 @matteobachetti
@warrickball I don't know how to pick up the OrcIDs, some people (me included) do have OrcID information on Github, but it wasn't pick up automatically. Is there any other step that needs to be taken?
