### 2024-10-24 #7389 @warrickball
> @warrickball I don't know how to pick up the OrcIDs, some people (me included) do have OrcID information on Github, but it wasn't pick up automatically. Is there any other step that needs to be taken?

Interesting! I don't know the process well but will see if I can find out more. It isn't an issue for the JOSS publication, so I'll advance with that anyway.
