### 2024-10-24 #7389 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
